Migrate git related tests
