+++ konf-core/src/test/kotlin/com/uchuhimo/konf/source/ProviderSpec.kt
@@ -141,52 +141,6 @@ object ProviderSpec : SubjectSpek<Provider>({
                     throws<SourceNotFoundException>())
             }
         }
-        /* TODO: Move to konf-jgit
-        on("create source from git repository") {
-            createTempDir().let { dir ->
-                Git.init().apply {
-                    setDirectory(dir)
-                }.call().use { git ->
-                    Paths.get(dir.path, "test").toFile().writeText("type = git")
-                    git.add().apply {
-                        addFilepattern("test")
-                    }.call()
-                    git.commit().apply {
-                        message = "init commit"
-                    }.call()
-                }
-                val repo = dir.toURI()
-                val source = subject.fromGit(repo.toString(), "test")
-                it("should create from the specified git repository") {
-                    assertThat(source.context["repo"], equalTo(repo.toString()))
-                    assertThat(source.context["file"], equalTo("test"))
-                    assertThat(source.context["branch"], equalTo(Constants.HEAD))
-                }
-                it("should return a source which contains value in git repository") {
-                    assertThat(source["type"].toText(), equalTo("git"))
-                }
-            }
-        }
-        on("create source from invalid git repository") {
-            createTempDir().let { dir ->
-                Git.init().apply {
-                    setDirectory(dir)
-                }.call().use { git ->
-                    Paths.get(dir.path, "test").toFile().writeText("type = git")
-                    git.add().apply {
-                        addFilepattern("test")
-                    }.call()
-                    git.commit().apply {
-                        message = "init commit"
-                    }.call()
-                }
-                it("should throw InvalidRemoteRepoException") {
-                    assertThat({ subject.fromGit(createTempDir().path, "test", dir = dir.path) },
-                        throws<InvalidRemoteRepoException>())
-                }
-            }
-        }
-        */
     }
 })
